# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Shared Assets APIs",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/register": {
            "post": {
                "tags": [
                    "Register"
                ],
                "summary": "Create admin",
                "operationId": "cac00d2dc7f1ea50d6ca5961b623c979",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "email": {
                                                "type": "string"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "password": {
                                                "type": "string"
                                            },
                                            "password_confirmation": {
                                                "type": "string"
                                            },
                                            "role": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "example@email.com",
                                    "name": "John Doe",
                                    "password": "123456",
                                    "password_confirmation": "123456",
                                    "role": 0
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "email"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "name"
                                        },
                                        "password": {
                                            "type": "string",
                                            "example": "password"
                                        },
                                        "role": {
                                            "type": "string",
                                            "example": 0
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2022-10-11T09:25:53.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2022-10-11T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/user/register": {
            "post": {
                "tags": [
                    "Register"
                ],
                "summary": "Admin creates a user",
                "operationId": "3bf6465bd98489a509343503a3b348d4",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "email": {
                                                "type": "string"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "password": {
                                                "type": "string"
                                            },
                                            "team_id": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "example@email.com",
                                    "name": "John Doe",
                                    "password": "123456",
                                    "team_id": 1
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "email"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "name"
                                        },
                                        "password": {
                                            "type": "string",
                                            "example": "password"
                                        },
                                        "role": {
                                            "type": "string",
                                            "example": 1
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2022-10-11T09:25:53.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2022-10-11T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "fail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}